Add support for Sereal as an input/output format
Add support for Sereal as an input/output format. We now have a hard
dependency on because it's awesome and you should use it instead of
using JSON.

The addition of Sereal support in the "serialize" subroutine mostly
involved copy/paste, the only original piece of code there is:

  if ($sereal_decoder_object->looks_like_sereal($uncompr)) {
    print STDERR "FlexSerializer: Detected that the input was Sereal\n" if DEBUG;
    print STDERR "FlexSerializer: This was the Sereal input: '%s'\n",
          substr($uncompr, 0, min(length($uncompr), 100)) if DEBUG >= 2;
    my $structure;
    $sereal_decoder_object->decode($uncompr, $structure);
    push @out, $structure;
  }

The rest is just copy/pasting to deal with us now needing to detect
some combination of Storable, JSON and Sereal. This patch is best
viewed with "git show -w" to make that easier to read.

Also do a major update to the documentation, with some changes to
SYNOPSIS that could have been split up into another patch, but it was
easier to make the change this way.

This also adds exhaustive testing for the new format, the only caveat
is that we're skipping some tests because under Sereal the serialized
string may not be the same, but the result will be, so we have these
TODO tests:

    $ perl -Ilib t/10basic.t | grep -C3 TODO | grep -a srl
    #          got: '=srl(*cbazCcfoocbar'
    #     expected: '=srlRcbazCcfoocbar'
    #          got: '=srl(*cbazCcfoocbar'
    #     expected: '=srlRcbazCcfoocbar'

They still deserialize to the same objects though, so this should be
OK. It might be hash ordering differences or the caveat that Serial
might choose to use the PV slot of an IV/PV where JSON/Storable will
pick both or the IV.