Fix regression in 1.09? that broke support for dynamically changing the output format
Adjust the test that I added in efc7d5e to indicate that changing
->output_format() on runtime no longer works.

I added support for this in efc7d5e, but it's been broken recently
with the patches that add the ability to add custom serializers. This
makes it much harder to do experiments with different output formats,
i.e. you have to construct two different serializers.

This resolves issue #5: https://github.com/tsee/Data-FlexSerializer/issues/5