Add tests to show that the things we advertise as rw aren't actually rw
Since 1.09 when we've had the magical pluggable formats we've been
failing horribly on cases where we did the right thing before,
i.e. when we change the input/output format at runtime.

I did a limited patch to solve the latter in the last commit, but we
still do completely the wrong thing with input formats. If you change
$flex->detect_*(Bool) it's just silently ignored.

Arguably we should just not support any rw operations at all. It makes
the implementation much more painful, but we shouldn't advertise that
it works and then completely fail on it.

So add failing tests to show that changing it at runtime completely
fails, we can then decide how to proceed with this. I.e. whether to
fix this up or to remove support for dynamically changing the output
format.