And it rises like a phoenix from the ashes
The whole support for changing the object after construction could
never work well, and was broken as noted in 28fde2f, 3893f08 and issue
#6.

So just remove support for it, and adjust the tests and the docs a
lot. This is basically a whole rewrite of the docs to take into
account how the API has changed since we initially wrote the module.